Spooky Bug Fix
Few things are spookier than finding a bug almost a year after you "finished" a project. I should've made that my costume tonight.
The bug was a fault in the binary searcher's edge case catcher, which caused it to effectively repeat the previous search without checking the other half. After seeing the time cost of that catcher when the fit is messier, I made it an option for the user (in userFunction.cpp).
While I was at it, I tidied up userFunction.cpp's partial derivatives section.